word_to_md/requirements.txt
python-docx>=0.8.11
transformers>=4.0.0